+++ docs/CHANGELOG.md
@@ -6,6 +6,7 @@ Changelog
 - Enh: Moved advanced page options into extra section
 - Enh: Compiled multiple translation messages into single `base` file
 - Fix #177: CLI error when no REST module is installed
+- Enh #148: Store folding state for categories per user
 
 
 1.5.3  (April 16, 2021)